+++ resources/js/bootstrap/store.js
@@ -1,4 +1,4 @@
-import { createStore } from 'vuex';
+import {createStore} from 'vuex';
 
 import {
     sortTree,
@@ -276,29 +276,51 @@ export const store = createStore({
                             // if no id exists, this data is added
                             if(!entity.data[k].id) {
                                 entity.data[k] = data.new_data[k];
+                                entity.data[k].value = data.data[k];
                                 if(data.sync) {
                                     state.entity.data[k] = data.new_data[k];
+                                    state.entity.data[k].value = data.data[k];
                                 }
                             } else {
-                                entity.data[k] = data.data[k];
+                                entity.data[k].value = data.data[k];
                                 if(data.sync) {
-                                    state.entity.data[k] = data.data[k];
+                                    state.entity.data[k].value = data.data[k];
                                 }
                             }
                         }
                     }
                 },
+                updateEntityMetadata(state, data) {
+                    let metadata = {};
+                    for(let k in data.data) {
+                        metadata[k] = data.data[k];
+                    }
+
+                    const eid = data.eid;
+                    if(!state.entities?.[eid]) {
+                        return;
+                    }
+
+                    if(state.entities?.[eid].metadata) {
+                        state.entity.metadata = {};
+                    }
+
+                    state.entities[eid].metadata = {
+                        ...state.entities[eid].metadata,
+                        ...metadata,
+                    };
+                },
                 updateEntityDataModerations(state, data) {
                     const entity = state.entities[data.entity_id];
-                    for(let i=0; i<data.attribute_ids.length; i++) {
+                    for(let i = 0; i < data.attribute_ids.length; i++) {
                         const curr = data.attribute_ids[i];
                         entity.data[curr].moderation_state = data.state;
                     }
                 },
                 addEntityTypeAttribute(state, data) {
                     const attrs = state.entityTypeAttributes[data.entity_type_id];
-                    attrs.splice(data.position-1, 0, data);
-                    for(let i=data.position; i<attrs.length; i++) {
+                    attrs.splice(data.position - 1, 0, data);
+                    for(let i = data.position; i < attrs.length; i++) {
                         if(attrs[i].position) {
                             attrs[i].position++;
                         } else if(attrs[i].pivot && attrs[i].pivot.position) {
@@ -312,7 +334,7 @@ export const store = createStore({
                     if(idx > -1) {
                         attrs.splice(idx, 1);
                     }
-                    for(let i=idx; i<attrs.length; i++) {
+                    for(let i = idx; i < attrs.length; i++) {
                         if(attrs[i].position) {
                             attrs[i].position++;
                         } else if(attrs[i].pivot && attrs[i].pivot.position) {
@@ -346,7 +368,7 @@ export const store = createStore({
                     const movedAttrs = attrs.splice(from, 1);
                     attrs.splice(to, 0, ...movedAttrs);
                     if(from < to) {
-                        for(let i=from; i<to; i++) {
+                        for(let i = from; i < to; i++) {
                             if(attrs[i].position) {
                                 attrs[i].position++;
                             } else if(attrs[i].pivot && attrs[i].pivot.position) {
@@ -354,7 +376,7 @@ export const store = createStore({
                             }
                         }
                     } else {
-                        for(let i=to+1; i<=from; i++) {
+                        for(let i = to + 1; i <= from; i++) {
                             if(attrs[i].position) {
                                 attrs[i].position--;
                             } else if(attrs[i].pivot && attrs[i].pivot.position) {
@@ -836,6 +858,9 @@ export const store = createStore({
                 },
                 updateEntityData({commit}, data) {
                     commit('updateEntityData', data);
+                },
+                updateEntityMetadata({commit}, data) {
+                    commit('updateEntityMetadata', data);
                 },
                 updateEntityDataModerations({commit}, data) {
                     commit('updateEntityDataModerations', data);