+++ resources/js/components/search/Simple.vue
@@ -21,6 +21,7 @@
         :delay="delay"
         :limit="limit"
         :placeholder="t('global.search')"
+        :disabled="disabled"
         @change="handleSelection"
     >
         <template #singlelabel="{ value }">
@@ -28,13 +29,16 @@
                 {{ displayResult(value) }}
             </div>
         </template>
-        <template #tag="{ option, handleTagRemove, disabled }">
-            <div class="multiselect-tag">
+        <template #tag="{ option, handleTagRemove, disabled: tagDisabled }">
+            <div
+                class="multiselect-tag"
+                :class="{'pe-2': tagDisabled}"
+            >
                 <span @click.prevent="handleTagClick(option)">
                     {{ displayResult(option) }}
                 </span>
                 <span
-                    v-if="!disabled"
+                    v-if="!tagDisabled"
                     class="multiselect-tag-remove"
                     @click.prevent
                     @mousedown.prevent.stop="handleTagRemove(option, $event)"
@@ -149,6 +153,11 @@
                 required: false,
                 default: null,
             },
+            disabled: {
+                type: Boolean,
+                required: false,
+                default: false,
+            },
         },
         emits: ['selected', 'entry-click'],
         setup(props, context) {
@@ -164,6 +173,7 @@
                 chain,
                 mode,
                 defaultValue,
+                disabled,
             } = toRefs(props);
             
             if(!keyText.value && !keyFn.value) {
@@ -258,5 +268,5 @@
                 state,
             };
         },
-    }
+    };
 </script>