+++ resources/js/components/tree/Node.vue
@@ -100,7 +100,7 @@
             data: {
                 required: true,
                 type: Object
-            }
+            },
         },
         setup(props) {
             const { t } = useI18n();