+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - Pasting into _Bibliography Item_ now works without clicking first and display some more info
 - Info on hover to required fields in new/edit bibliography modal
 - Restrict search in _Entity_ attribute to specific entity-types
+- Metadata tab to _Entity Detail_ panel=======
 - Allow _Multiple Choice Dropdown_ in _Table_ attribute
 - Allow _Date Range_ in _Table_ attribute
 - Entities can now be updated using the _Data Importer_
@@ -92,6 +93,10 @@ All notable changes to this project will be documented in this file.
 - Append `et al.` to citations in reference tab and modal
 - Do not show "!" for attributes without value & replace initial "!" with "?" for attributes with value, but no certainty set
 - Bibtex import now is case-insensitive for type and keys
+- Display only the first and last 10 rows of a table that has more than 25 rows
+  - Click on _show hidden_ text to display all rows
+- Only load first 15 elements in dropdown attributes and load more elements on scroll/search (due to bad performance if an entity has a lot of dropdowns)
+- Reworked attribute list logic for a more consistent data management in the background
 
 ## 0.9.14
 ### Added